requirements.txt
google-auth==2.4.1
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.16.0
python-dotenv==0.19.1
python-telegram-bot==13.7
